refactor(control-mode): extract shared capture list + replay assembly (PaneCaptureReplay)
Pure extraction, no behavior change: the 6-command capture list (with its
live-probed leg semantics) and the parse/truncation-telemetry/leg-
recombination/assemble body of replayAndOpenGate move to PaneCaptureReplay,
and the orchestrator's private BatchCollector/sendBatch become a shared
TmuxControlCommandClient.sendBatch(texts:). The Phase B M3 overflow-repair
coordinator reuses both — its repair replay must stay byte-identical to the
attach's.
